+++ serialization-msgpack/src/commonTest/kotlin/com/ensarsarajcic/kotlinx/serialization/msgpack/MsgPackTest.kt
@@ -5,14 +5,18 @@ import kotlinx.serialization.KSerializer
 import kotlinx.serialization.builtins.ArraySerializer
 import kotlinx.serialization.builtins.ByteArraySerializer
 import kotlinx.serialization.builtins.MapSerializer
+import kotlinx.serialization.builtins.PairSerializer
+import kotlinx.serialization.builtins.TripleSerializer
 import kotlinx.serialization.builtins.nullable
 import kotlinx.serialization.builtins.serializer
+import kotlinx.serialization.decodeFromByteArray
 import kotlinx.serialization.encodeToByteArray
-import kotlinx.serialization.serializer
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 import kotlin.test.fail
+
+typealias NestedMessage = Pair<List<Pair<String, String>>, String>
 
 internal class MsgPackTest {
     @Test
@@ -288,6 +292,41 @@ internal class MsgPackTest {
         testPairs(TestData.uByteTestPairs.map { it.first }, Byte.serializer())
         testPairs(TestData.uShortTestPairs.map { it.first }, Short.serializer())
         testPairs(TestData.uIntTestPairs.map { it.first }, Int.serializer())
+    }
+
+    @Test
+    fun testPairsEncode() {
+        testEncodePairs(PairSerializer(String.serializer(), String.serializer()), *TestData.pairsTestPairs)
+    }
+
+    @Test
+    fun testPairsDecode() {
+        testDecodePairs(PairSerializer(String.serializer(), String.serializer()), *TestData.pairsTestPairs)
+    }
+
+    @Test
+    fun testTriplesEncode() {
+        testEncodePairs(
+            TripleSerializer(String.serializer(), String.serializer(), String.serializer()),
+            *TestData.triplesTestPairs
+        )
+    }
+
+    @Test
+    fun testTriplesDecode() {
+        testDecodePairs(TripleSerializer(String.serializer(), String.serializer(), String.serializer()), *TestData.triplesTestPairs)
+    }
+
+    @Test
+    fun testNestedStructures() {
+        val sm1: NestedMessage = listOf("Alice" to "Bob", "Charley" to "Delta") to "Random message Body here"
+        val result = MsgPack.encodeToByteArray(sm1)
+        println(result.toHex())
+        println(result.toList())
+        println(result.size)
+        println(MsgPack.decodeFromByteArray<Pair<Int, Int>>(MsgPack.encodeToByteArray(1 to 2)))
+        val result2 = MsgPack.decodeFromByteArray<NestedMessage>(result)
+        assertEquals(sm1, result2)
     }
 
     @Test