+++ serialization-msgpack/src/commonTest/kotlin/com/ensarsarajcic/kotlinx/serialization/msgpack/TestData.kt
@@ -149,6 +149,12 @@ object TestData {
     val sampleClassTestPairs = arrayOf(
         "83aa74657374537472696e67a3646566a774657374496e747bab74657374426f6f6c65616ec3" to SampleClass("def", 123, true)
     )
+    val pairsTestPairs: Array<Pair<String, Pair<String, String>>> = arrayOf(
+        "82a56669727374a5416c696365a67365636f6e64a3426f62" to Pair("Alice", "Bob")
+    )
+    val triplesTestPairs: Array<Pair<String, Triple<String, String, String>>> = arrayOf(
+        "83a56669727374a5416c696365a67365636f6e64a3426f62a57468697264a454657374" to Triple("Alice", "Bob", "Test")
+    )
 }
 
 @Serializable(with = CustomExtensionSerializer::class)