+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+- Fixed issues with nested structure deserialization ([#57][i57])
 
 ## [0.4.1] - 2021-10-24
 - Fixed issues with platform specific release artifacts ([#55][i55])
@@ -86,4 +87,5 @@ MsgPack.default.encodeToByteArray(...)
 [i19]: https://github.com/esensar/kotlinx-serialization-msgpack/issues/19
 [i20]: https://github.com/esensar/kotlinx-serialization-msgpack/issues/20
 [i55]: https://github.com/esensar/kotlinx-serialization-msgpack/issues/55
+[i57]: https://github.com/esensar/kotlinx-serialization-msgpack/issues/57
 [p40]: https://github.com/esensar/kotlinx-serialization-msgpack/pull/40